+++ libs/bpm-core-client/src/lib/form/index.ts
@@ -1,4 +1,5 @@
 export * from './form-api';
 export * from './form-data-source-api';
+export * from './form-data-source-builder';
 export * from './form-data-source-state';
 export * from './form-rendering';